pyproject.toml
[project]
name = "product-management"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Marcus Floriano", email = "marcus.floriano@gmail.com" }
]
dependencies = [
    "django==5.1",
    "django-allauth[socialaccount]>=65.0.2",
    "django-debug-toolbar>=4.4.6",
    "django-quill-editor>=0.1.42",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.rye.scripts]
runserver = "python src/manage.py runserver"
